test: add e2e test
