+++ packages/plugins/@nocobase/plugin-action-duplicate/src/client/__e2e__/schemaSettings.test.ts
@@ -8,7 +8,7 @@
  */
 
 import { expect, test } from '@nocobase/test/e2e';
-import { oneEmptyTableBlockWithDuplicateActions } from './utils';
+import { T4546, oneEmptyTableBlockWithDuplicateActions } from './templates';
 
 test.describe('direct duplicate & copy into the form and continue to fill in', () => {
   test('direct duplicate', async ({ page, mockPage, mockRecords }) => {
@@ -95,5 +95,46 @@ test.describe('direct duplicate & copy into the form and continue to fill in', (
     expect(postData.manyToOne['id']).toBe(data.manyToOne['id']);
     expect(postData.oneToOneBelongsTo['id']).toBe(data.oneToOneBelongsTo['id']);
     expect(manyToMany).toEqual(expect.arrayContaining(expectManyToMany));
+  });
+
+  test('association block in popup', async ({ page, mockPage, mockRecord }) => {
+    const nocoPage = await mockPage(T4546).waitForInit();
+    const general1 = await mockRecord('general1');
+    await nocoPage.goto();
+
+    // 1. 先打开弹窗，然后点击弹窗中复制按钮，应该能成功复制，并显示在弹窗中
+    await page.getByLabel('action-Action.Link-View record-view-general1-table-0').click();
+    await page.getByLabel('action-Action.Link-Duplicate-duplicate-general2-table-0').click();
+    await expect(
+      page.getByLabel('block-item-CardItem-general2-').getByText(general1.oneToMany[0].singleLineText2),
+    ).toHaveCount(2);
+
+    // 2. 设置复制方式为编辑模式，点击复制按钮，应该能成功复制，并显示在弹窗中
+    await page.getByLabel('action-Action.Link-Duplicate-duplicate-general2-table-0').hover();
+    await page
+      .getByRole('button', { name: 'designer-schema-settings-Action.Link-actionSettings:duplicate-general2' })
+      .hover();
+    await page.getByRole('menuitem', { name: 'Duplicate mode' }).click();
+    await page.getByLabel('Copy into the form and').check();
+    await page.getByRole('button', { name: 'OK', exact: true }).click();
+
+    // 设置表单
+    await page.getByLabel('action-Action.Link-Duplicate-duplicate-general2-table-0').click();
+    await page
+      .getByTestId('drawer-Action.Container-general2-Duplicate')
+      .getByLabel('schema-initializer-Grid-popup')
+      .hover();
+    await page.getByRole('menuitem', { name: 'form Form right' }).hover();
+    await page.getByRole('menuitem', { name: 'Current collection' }).click();
+    await page
+      .getByTestId('drawer-Action.Container-general2-Duplicate')
+      .getByLabel('schema-initializer-ActionBar-')
+      .hover();
+    await page.getByRole('menuitem', { name: 'Submit' }).click();
+    await page.getByLabel('action-Action-Submit-submit-').click();
+
+    await expect(
+      page.getByLabel('block-item-CardItem-general2-').getByText(general1.oneToMany[0].singleLineText2),
+    ).toHaveCount(3);
   });
 });
